# Block asserts into switch statements 