Moving buffer asserts into switch statements  blocks seems significantly speed up register allocation and coalescing. This reduces compile opt-pass time for resnet50 (autoschedule_with_convnets branch) from 268.5848 sec to 15 sec. 

$ run_performance_lots_of_inputs 
Before:
```
Compile time with 1 inputs = 0.534758 s
Compile time with 2 inputs = 0.078090 s
Compile time with 4 inputs = 0.113272 s                                                                            
Compile time with 8 inputs = 0.194721 s                                                                           
Compile time with 16 inputs = 0.401624 s                                                                             
Compile time with 32 inputs = 0.940241 s                                                                          
Compile time with 64 inputs = 2.707179 s                                                                             
Compile time with 128 inputs = 8.599611 s

```
$ run_performance_lots_of_inputs
After:
```
Compile time with 1 inputs = 0.523519 s
Compile time with 2 inputs = 0.083440 s
Compile time with 4 inputs = 0.125427 s
Compile time with 8 inputs = 0.258036 s
Compile time with 16 inputs = 0.487263 s
Compile time with 32 inputs = 1.036269 s
Compile time with 64 inputs = 2.339884 s
Compile time with 128 inputs = 5.862377 s
```

Resnet50 compile time before : 

```
===-------------------------------------------------------------------------===                                       
                      ... Pass execution timing report ...                                                             
===-------------------------------------------------------------------------===                                        
  Total Execution Time: 268.5848 seconds (270.2123 wall clock)                                                         
                                                                                                                       
   ---User Time---   --System Time--   --User+System--   ---Wall Time---  --- Name ---                                 
  224.7175 ( 84.1%)   0.6710 ( 50.7%)  225.3886 ( 83.9%)  226.9213 ( 84.0%)  Simple Register Coalescing                
  15.1290 (  5.7%)   0.2711 ( 20.5%)  15.4001 (  5.7%)  15.4697 (  5.7%)  Greedy Register Allocator


```
Resnet50 compile time after : 


```
===-------------------------------------------------------------------------===
                      ... Pass execution timing report ...
===-------------------------------------------------------------------------===
  Total Execution Time: 15.8371 seconds (15.8704 wall clock)

   ---User Time---   --System Time--   --User+System--   ---Wall Time---  --- Name ---
   1.9074 ( 12.2%)   0.0634 ( 24.9%)   1.9708 ( 12.4%)   1.9764 ( 12.5%)  Greedy Register Allocator

```